Move wrap_route function
