+++ qcfractal/qcfractal/components/auth/routes.py
@@ -4,7 +4,7 @@ from flask import current_app, g
 
 from qcfractal.flask_app import storage_socket
 from qcfractal.flask_app.api_v1.blueprint import api_v1
-from qcfractal.flask_app.api_v1.helpers import wrap_route
+from qcfractal.flask_app.wrap_route import wrap_route
 from qcportal.auth import UserInfo, RoleInfo, GroupInfo
 from qcportal.exceptions import (
     InconsistentUpdateError,
