+++ qcfractal/qcfractal/components/managers/routes.py
@@ -2,8 +2,8 @@ from flask import current_app
 
 from qcfractal.flask_app import storage_socket
 from qcfractal.flask_app.api_v1.blueprint import api_v1
-from qcfractal.flask_app.api_v1.helpers import wrap_route
 from qcfractal.flask_app.compute_v1.blueprint import compute_v1
+from qcfractal.flask_app.wrap_route import wrap_route
 from qcportal.base_models import CommonBulkGetNamesBody
 from qcportal.exceptions import LimitExceededError
 from qcportal.managers import (
