+++ qcfractal/qcfractal/components/reaction/routes.py
@@ -4,7 +4,7 @@ from flask import current_app, g
 
 from qcfractal.flask_app import storage_socket
 from qcfractal.flask_app.api_v1.blueprint import api_v1
-from qcfractal.flask_app.api_v1.helpers import wrap_route
+from qcfractal.flask_app.wrap_route import wrap_route
 from qcportal.exceptions import LimitExceededError
 from qcportal.reaction import (
     ReactionDatasetSpecification,
