+++ qcfractal/qcfractal/flask_app/compute_v1/routes.py
@@ -1,6 +1,6 @@
-from qcfractal.flask_app.api_v1.helpers import wrap_route
 from qcfractal.flask_app.compute_v1.blueprint import compute_v1
 from qcfractal.flask_app.helpers import get_public_server_information
+from qcfractal.flask_app.wrap_route import wrap_route
 
 
 @compute_v1.route("/information", methods=["GET"])
